fixed: a bug about ObjectInstance.world format
Previously the format attribute for ObjectInstance.world was set to eR32G32B32Sfloat.
This mistake caused the last row of each instance's world transform to be [1, 1, 1, 1], not [0, 0, 0, 1],
in which case the w-coordinate after world transformation was set to non-one.
The correct format is eR32G32B32A32Sfloat.

Additionally refactored vertex input attribute code.